// src/app/models/count.ts
export class Count {
   an: number;
   ap: number;
   ar: number;
   as: number;
   br: number;
   ch: number;
   ct: number;
   date: string;
   dd: number;
   dl: number;
   dn: number;
   ga: number;
   gj: number;
   hp: number;
   hr: number;
   jh: number;
   jk: number;
   ka: number;
   kl: number;
   la: number;
   ld: number;
   mh: number;
   ml: number;
   mn: number;
   mp: number;
   mz: number;
   nl: number;
   or: number;
   pb: number;
   py: number;
   rj: number;
   sk: number;
   status: string;
   tg: number;
   tn: number;
   tr: number;
   tt: number;
   un: number;
   up: number;
   ut: number;
   wb: number
}