// src/app/models/state-tests.ts
export class StateTests {
   coronaenquirycalls: number;
   cumulativepeopleinquarantine: number;
   negative: number;
   numcallsstatehelpline: number;
   numicubeds: number;
   numisolationbeds: number;
   numventilators: number;
   peopleinicu: number;
   peopleonventilators: number;
   populationncp2019projection: number;
   positive: number;
   source1: string;
   source2: string;
   state: string;
   tagpeopleinquarantine: string;
   tagtotaltested: string;
   testpositivityrate: number;
   testspermillion: number;
   testsperpositivecase: number;
   testsperthousand: number;
   totaln95masks: number;
   totalpeoplecurrentlyinquarantine: number;
   totalpeoplereleasedfromquarantine: number;
   totalppe: number;
   totaltested: number;
   unconfirmed: number;
   updatedon: string;
}